Unify paused states and add ConfirmReset state
I totally missed the ConfirmReset UI state, so that had to be added. I also
realized that the two paused states are visually identical, and as this state
machine is for the UI, they can be combined. Tracking whose turn it is should
happen in the data model, not the UI state machine.