Add UI updating to Game with CADisplayLink
Previously we updated the screen with an NSTimer in the view controller.
It did the job, but the more natural means of frequently updating the
views -- particularly the players' times -- is with a CADisplayLink,
occurring at (up to) every frame.

With the current view design, displaying at 1-second resolution, this is
overkill; we don't need to update 60 times per second. But by
implementing this now, there is flexibility for design changes in the
future. Also it's just more natural and fits better than NSTimer.