Finish implementing Pause state logic
With the new approach to ownership of interface changes (i.e. each state
is responsible for undoing in the willExit method what they did in the
didEnter method), dealing with the player buttons and the clocks is
unnecessary. We just need to change the pause and reset buttons.

Dealing with the transitions from the Loading state and the ConfirmReset
state is a bit repetitive. It's the best I can do until I move
transition logic into the DMGameView.