Add pauseGame method to VC for AppDelegate to call
When we resign active, the AppDelegate has to pause the game. This could
be exposed through the .game property on the VC, but then the
AppDelegate would have to know what State transition the Game into, and
how to do that. Since the VC mediates between the outside world already
(touch events), it makes sense that it would be responsible for
mediating this outside world event.