Fix white slider constraint referencing outlet
Ughhhhhhhhhhh. This took too long to find.

The outlet for the white slider top constraint was assigned to the top
constraint for the startGameLabel. So the label was never visible, since
that constraint was told to be -1 in -setUpSubviews. (And as such it
never would be, regardless of changes made in the storyboard, because
it was modified at runtime.) That was a costly misclick. Note to self:
always look for this in the future.