Add actions for slider end editing
This avoids updating the model a huge number of times while dragging the
slider.

The sliders' TouchUpInside and TouchUpOutside events are both hooked up
to the respective _____FinishedEditing methods on the view.

Amusing aside: in figuring out which event to use, I discovered that my
old code had a bug: it didn't do anything on TouchUpOutside. So if the
user did that, the clock view would be updated based on the slider view,
but then when they started the game, their clock would be set at
something else.