Rewrite README completely, and add related images
The earlier stab at a README was getting really wordy and unwieldy, so
this trims all the fat and leaves a (hopefully) simple, straightforward
piece of prose.

It also includes a composite screenshot image, and the class and state
diagrams. I should probably update the diagrams as they have problems,
but I'd rather push this now and fix that later.